+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased]
+## 1.1.0 - 2016-03-01
 - Added support for `-n <number>` or `--definitions=<number>` to limit output
 - Fixed output mangled by carriage returns (`\r`)
 - Added `--format=definition` that only outputs the definitions for a word